requirements.txt
# Requirements for EXE build
# python==3.10.6
sympy==1.11.1
numpy==1.24.1
matplotlib==3.6.2
pyyaml==6.0
markdown==3.4.1
pint==0.20.1
pyqt5==5.15.7
scipy==1.10.0
PyInstaller==5.7  # For building exe
pip-licenses      # Needed to generate license info for exe about page
pytest            # To run test cases
